algo-bot/backend/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
requests>=2.31.0
tenacity>=8.0.0